P3 create tags table and model

See merge request seng302-2019/team-400!178

+++ app/repository/PhotoRepository.java
@@ -1,21 +1,18 @@
 package repository;
 
 import static java.util.concurrent.CompletableFuture.supplyAsync;
-import static play.mvc.Results.ok;
 
 import com.google.common.collect.Iterables;
 import io.ebean.Ebean;
 import io.ebean.EbeanServer;
 import java.util.Collection;
 import java.util.List;
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import models.DestinationPhoto;
 import models.Photo;
 import play.db.ebean.EbeanConfig;
-import play.mvc.Result;
 import util.objects.Pair;
 
 /**
@@ -24,7 +21,7 @@ import util.objects.Pair;
 @Singleton
 public class PhotoRepository {
 
-    public static final String FRONTEND_APPEND_DIRECTORY = "../user_content/";
+    private static final String FRONTEND_APPEND_DIRECTORY = "../user_content/";
     private static final String USER_ID = "user_id";
     private static final String IS_PROFILE = "is_profile";
     private final EbeanServer ebeanServer;
@@ -42,10 +39,10 @@ public class PhotoRepository {
      * @param photo Photo to be added
      * @return Ok on success
      */
-    public CompletableFuture<Result> addPhoto(Photo photo) {
+    public CompletableFuture<Long> addPhoto(Photo photo) {
         return supplyAsync(() -> {
             ebeanServer.insert(photo);
-            return ok();
+            return photo.guid;
         }, executionContext);
     }
 
@@ -53,7 +50,8 @@ public class PhotoRepository {
      * Clears any existing profile photo for a user, and returns filenames of old files.
      *
      * @param userID user ID to clear profile photo of
-     * @return OK if successfully cleared existing profile pic, notFound if none found
+     * @return A Pair of strings, the key being the filename of the photo and the value being the
+     * thumbnail filename of the photo. Returns null if the user didn't have a profile photo
      */
     public CompletableFuture<Pair<String, String>> clearProfilePhoto(long userID) {
         return supplyAsync(() -> {
@@ -74,7 +72,7 @@ public class PhotoRepository {
     }
 
     /**
-     * Finds all photos in database related to the given user ID.
+     * Finds all photos in database related to the given user ID that aren't a profile photo
      *
      * @param userID User to find all photos for
      * @return List of Photo objects with the specified user ID
@@ -92,7 +90,7 @@ public class PhotoRepository {
     }
 
     /**
-     * Finds all public photos in database related to the given user ID.
+     * Finds all public photos in database related to the given user ID that aren't a profile photo
      *
      * @param userID User to find all public photos for
      * @return List of Photo objects with the specified user ID
@@ -114,7 +112,8 @@ public class PhotoRepository {
      * Finds the profile picture in the database for the given user ID.
      *
      * @param userID User to find profile picture for
-     * @return a photo
+     * @return a photo, which will be null if they user doesn't exist or doesn't have a profile
+     * picture
      */
     public CompletableFuture<Photo> getUserProfilePicture(Long userID) {
         return supplyAsync(() -> {
@@ -156,7 +155,7 @@ public class PhotoRepository {
      * Deletes a photo from the database.
      *
      * @param id Unique photo ID of destination to be deleted
-     * @return The number of rows deleted
+     * @return The deleted photo, or null if a photo with that id was not found
      */
     public CompletableFuture<Photo> deletePhoto(Long id) {
         return supplyAsync(() -> {
@@ -166,7 +165,9 @@ public class PhotoRepository {
                     .findOneOrEmpty()
                     .orElse(null);
 
-                photo.delete();
+                if (photo != null) {
+                    photo.delete();
+                }
                 return photo;
             }
             , executionContext);
@@ -204,21 +205,7 @@ public class PhotoRepository {
      * Get photo object from db where it has some filename.
      *
      * @param filename Name of file
-     */
-    public CompletableFuture<Photo> getPhotoByFilename(String filename) {
-        return supplyAsync(() -> ebeanServer.find(Photo.class)
-                .where()
-                .eq("filename", filename)
-                .findOneOrEmpty()
-                .orElse(null),
-            executionContext);
-
-    }
-
-    /**
-     * Get photo object from db where it has some filename.
-     *
-     * @param filename Name of file
+     * @return true if deleted, false if not
      */
     public CompletableFuture<Boolean> deletePhotoByFilename(String filename) {
         return supplyAsync(() -> {
@@ -226,7 +213,7 @@ public class PhotoRepository {
                 .where()
                 .eq("filename", filename)
                 .findOne();
-            if(photo != null) {
+            if (photo != null) {
                 photo.delete();
                 return true;
             } else {
@@ -239,6 +226,7 @@ public class PhotoRepository {
      * Update photo row in database.
      *
      * @param photo photo object to update
+     * @return the updated photo's guid
      */
     public CompletableFuture<Long> updatePhoto(Photo photo) {
         return supplyAsync(() -> {
@@ -248,6 +236,13 @@ public class PhotoRepository {
             executionContext);
     }
 
+    /**
+     * Gets a deleted destination photo form the database
+     *
+     * @param photoId The id of the photo to retrieve
+     * @param destId The id of the associated destination
+     * @return The destination photo, or null if not found
+     */
     public CompletableFuture<DestinationPhoto> getDeletedDestPhoto(Long photoId, Long destId) {
         return supplyAsync(() -> ebeanServer.find(DestinationPhoto.class)
             .setIncludeSoftDeletes()
@@ -256,5 +251,5 @@ public class PhotoRepository {
             .eq("destination_id", destId)
             .findOneOrEmpty()
             .orElse(null), executionContext);
-    };
+    }
 }
